requirements.txt
fastapi
uvicorn
weaviate-client
transformers
requests
